update case where no branch condition is met
